third_party/WebKit/LayoutTests/virtual/android/fullscreen/compositor-touch-hit-rects-fullscreen-video-controls-expected.txt

This test makes sure that touch hit rects are reported for fullscreen HTML5 video control elements even when there is a document handler.

Should have single rect on document before fullscreen
handler: #document (0, 0, 800, 600)

EVENT(webkitfullscreenchange)
Should report no rect or another rect which is not on the document
handler: DIV (42, 3, 637, 24)

END OF TEST
